adds new, modular handling of individual options
fixes #30
this is a massive restructuring of the options
each option gets its own function
each options tab has a separate function
and the tabs themselves are in a function
this means that you can rewrite the options page whilst still being
able to include options from museum-core in child themes without a lot
of duplication of code and without having to undo sections of code from
museum-core.

for example, if you wanted to write an alternate options page that
included some options from core you could do something like:

     function my_options_page() {
           ap_core_favicon_option();
           ap_core_css_option();
           my_cool_option();
     }

moving forward, I might further abstract the theme options page so the
layout stuff is all in its own functions so you would just need to run
a series of functions and not have to worry about the layout at all.