renames template parts that run the loop in the template part
template parts that contain a loop are handled in a different way then template parts in which the loop is run outside the template part (like parts for post formats). renaming these template parts to establish that these are inherently different types of template parts from the post formats. it was confusing anyway when going between one template part to the next when they bounce back and forth between templates that have a loop and templates that don't.